# application/services/custom_services/types/src/models/entitymodels/types.ts
export interface types 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: String,
   description: String
}